Module type Writer.S

type t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val close : t -> unit Async_kernel.Deferred.t

val is_closed : t -> bool

val monitor : t -> Async_kernel.Monitor.t

val bytes_to_write : t -> int

val stopped : t -> unit Async_kernel.Deferred.t

Becomes determined when it is no longer possible to send message using this writer, for instance when the writer is closed or the consumer leaves.

The result of this function is cached by Rpc.Transport

val flushed : t -> unit Async_kernel.Deferred.t

flushed t returns a deferred that must become determined when all prior sent messages are delivered.

It must be OK to call flushed t after t has been closed.

val ready_to_write : t -> unit Async_kernel.Deferred.t

ready_to_write t becomes determined when it is a good time to send messages again. Async RPC calls this function after sending a batch of messages, to avoid flooding the transport.

Using let ready_to_write = flushed is an acceptable implementation.

All the following functions send exactly one message.

val send_bin_prot : t -> 'a Core_kernel.Bin_prot.Type_class.writer -> 'a -> unit Async_rpc_kernel__Transport_intf.Send_result.t

val send_bin_prot_and_bigstring : t -> 'a Core_kernel.Bin_prot.Type_class.writer -> 'a -> buf:Core_kernel.Bigstring.t -> pos:int -> len:int -> unit Async_rpc_kernel__Transport_intf.Send_result.t

val send_bin_prot_and_bigstring_non_copying : t -> 'a Core_kernel.Bin_prot.Type_class.writer -> 'a -> buf:Core_kernel.Bigstring.t -> pos:int -> len:int -> unit Async_kernel.Deferred.t Async_rpc_kernel__Transport_intf.Send_result.t

Same as send_bin_prot_and_bigstring but the bigstring can't be modified until the returned deferred becomes determined. This can be used to avoid copying the bigstring.