Module Connection.Heartbeat_config

type t

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Core_kernel.Bin_prot.Binable.S with type t := t

include Core_kernel.Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val create : ?timeout:Core_kernel.Time_ns.Span.t -> ?send_every:Core_kernel.Time_ns.Span.t -> unit -> t

Each side of the connection has its own heartbeat config. It sends a heartbeat every send_every. If it doesn't receive a heartbeat for timeout, it drops the connection. It only checks whether timeout has elapsed when it sends heartbeats, so effectively timeout is rounded up to the nearest multiple of send_every.

val timeout : t -> Core_kernel.Time_ns.Span.t

val send_every : t -> Core_kernel.Time_ns.Span.t