Module Rpc.On_exception

When your implementation raises an exception, that exception might happen before a value is returned or after a value is returned. The latter kind of exception is what the ~rest flag to Monitor.try_with is dealing with.

In both these cases, callback will be called with the exception.

In the case where the exception raises before a value is returned, close_connection_if_no_return_value determines whether the connection is closed. Exceptions raised after a value is returned will never close a connection.

type t = {

callback : ( exn -> unit ) option;

close_connection_if_no_return_value : bool;

}

val close_connection : t

val continue : t