Module Rpc.Pipe_close_reason

type t =

| Closed_locally	(*
The RPC implementer closed the pipe.
*)

| Closed_remotely	(*
An error occurred, e.g. a message could not be deserialized. If the connection closes before either side explicitly closes the pipe, it will also go into this case.
*)

| Error of Core_kernel.Error.t

include Core_kernel.Bin_prot.Binable.S with type t := t

include Core_kernel.Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val compare : t -> t -> int

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

module Stable : sig ... end