Module Rpc.Transport

RPC transport layer

A transport is a way to send and receive messages. It is split between a Reader and a Writer part.

Async_rpc_kernel only provides an async Pipe.t transport. Async_rpc provides a unix transport based on Async_unix's Reader and Writer modules.

module Header : sig ... end

Binary headers containing message lengths. All transports should use this to ensure binary compatibility.

module Handler_result : sig ... end

module Reader : sig ... end

module Send_result : sig ... end

module Writer : sig ... end

type t = {

reader : Reader.t;

writer : Writer.t;

}

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val close : t -> unit Async_kernel.Deferred.t

Closes both parts of the transport.