Module State_rpc.Make

Parameters

module Model : sig ... end

Signature

module type Version_shared = sig ... end

module Register : functor (Version_i : sig ... end) -> sig ... end

add a new version to the set of versions available via implement_multi.

module Register_raw : functor (Version_i : sig ... end) -> sig ... end

Register_raw is like Register except you get the whole update pipe to deal with. This is useful if, e.g., your update_of_model function can fail, so that you'd like to filter items out from the result pipe.

include S with type query := Model.query with type state := Model.state with type update := Model.update with type error := Model.error

val implement_multi : ?⁠log_not_previously_seen_version:(name:string -> int -> unit) -> ('connection_state -> version:int -> Model.query -> (Model.state * Model.update Async_kernel.Pipe.Reader.t, Model.error) Core_kernel.Result.t Async_kernel.Deferred.t) -> 'connection_state Rpc.Implementation.t list

implement multiple versions at once

val rpcs : unit -> Rpc.Any.t list

All rpcs implemented by implement_multi

val versions : unit -> Core_kernel.Int.Set.t

All versions supported by implement_multi. (useful for computing which old versions may be pruned)

val name : string