Module Caller_converts.One_way

module type S = sig ... end

signature for rpc dispatchers

module Make : functor (Model : sig ... end) -> sig ... end

Given a model of the types involved in a family of RPCs, this functor provides a single RPC versioned dispatch function dispatch_multi in terms of that model and a mechanism for registering the individual versions that dispatch_multi knows about. Registration requires knowing how to get out of the model.