Module One_way.Make

Parameters

module Model : sig ... end

Signature

module Register : functor (Version_i : sig ... end) -> sig ... end

add a new version to the set of versions available via dispatch_multi.

include S with type msg := Model.msg

val dispatch_multi : Connection_with_menu.t -> Model.msg -> unit Core_kernel.Or_error.t

multi-version dispatch

val rpcs : unit -> Rpc.Any.t list

All rpcs supported by dispatch_multi

val versions : unit -> Core_kernel.Int.Set.t

All versions supported by dispatch_multi. (useful for computing which old versions may be pruned)

val name : string