Module type State_rpc.S

type query

type state

type update

type error

val dispatch_multi : Connection_with_menu.t -> query -> (state * update Core_kernel.Or_error.t Async_kernel.Pipe.Reader.t * Rpc.State_rpc.Metadata.t, error) Core_kernel.Result.t Core_kernel.Or_error.t Async_kernel.Deferred.t

Multi-version dispatch

The return type varies slightly from Rpc.State_rpc.dispatch to make it clear that conversion of each individual element in the returned pipe may fail.

val rpcs : unit -> Rpc.Any.t list

All rpcs supported by dispatch_multi

val versions : unit -> Core_kernel.Int.Set.t

All versions supported by dispatch_multi (useful for computing which old versions may be pruned).

val name : string