Module Async_rpc_kernel__Transport

RPC transport layer

A transport is a way to send and receive messages. It is split between a Reader and a Writer part.

Async_rpc_kernel only provides an async Pipe.t transport. Async_rpc provides a unix transport based on Async_unix's Reader and Writer modules.