Module Rpc.One_way

An RPC that has no response. Error handling is trickier here than it is for RPCs with responses, as there is no reasonable place to put an error if something goes wrong. Because of this, in the event of an error such as dispatching to an unimplemented RPC, the connection will be shut down. Similarly, if the implementation raises an exception, the connection will be shut down.

type 'msg t

val create : name:string -> version:int -> bin_msg:'msg Bin_prot.Type_class.t -> 'msg t

val name : _ t -> string

val version : _ t -> int

val description : _ t -> Description.t

val bin_msg : 'msg t -> 'msg Bin_prot.Type_class.t

val implement : 'msg t -> ('connection_state -> 'msg -> unit) -> 'connection_state Implementation.t

val dispatch' : 'msg t -> Connection.t -> 'msg -> unit Rpc_result.t

dispatch' exposes Rpc_result.t as output. Passing it through rpc_result_to_or_error gives you the same result as dispatch

val rpc_result_to_or_error : 'msg t -> Connection.t -> unit Rpc_result.t -> unit Core_kernel.Or_error.t

val dispatch : 'msg t -> Connection.t -> 'msg -> unit Core_kernel.Or_error.t

val dispatch_exn : 'msg t -> Connection.t -> 'msg -> unit

module Expert : sig ... end