Module One_way.Expert

val implement : _ t -> ('connection_state -> Core_kernel.Bigstring.t -> pos:int -> len:int -> unit) -> 'connection_state Implementation.t

val dispatch : _ t -> Connection.t -> Core_kernel.Bigstring.t -> pos:int -> len:int -> [ `Ok | `Connection_closed ]

val schedule_dispatch : _ t -> Connection.t -> Core_kernel.Bigstring.t -> pos:int -> len:int -> [ `Flushed of unit Async_kernel.Deferred.t | `Connection_closed ]

Like dispatch, but does not copy data out of the buffer, so it must not change until the returned unit Deferred.t is determined.