Module Group.Buffer

A group internally holds a buffer to serialize messages only once. This buffer will grow automatically to accomodate bigger messages.

It is safe to share the same buffer between multiple groups.

type t

val create : ?⁠initial_size:int -> unit -> t