Module One_way.Make

Parameters

module Model : sig ... end

Signature

module Register : functor (Version_i : sig ... end) -> sig ... end

Add a new version to the set of versions implemented by implement_multi.

include S with type msg := Model.msg

val implement_multi : ?⁠log_not_previously_seen_version:(name:string -> int -> unit) -> ('state -> version:int -> Model.msg -> unit) -> 'state Rpc.Implementation.t list

implement multiple versions at once

val rpcs : unit -> Rpc.Any.t list

All rpcs implemented by implement_multi

val versions : unit -> Core_kernel.Int.Set.t

All versions implemented by implement_multi. (useful for computing which old versions may be pruned)

val name : string