Module Rpc.Make

Parameters

module Model : sig ... end

Signature

module Register : functor (Version_i : sig ... end) -> sig ... end

Adds a new version to the set of versions available via dispatch_multi.

module Register' : functor (Version_i : sig ... end) -> sig ... end

A variant of Register in which the query is made available when transforming the response

include S with type query := Model.query with type response := Model.response

val dispatch_multi : Connection_with_menu.t -> Model.query -> Model.response Core_kernel.Or_error.t Async_kernel.Deferred.t

Multi-version dispatch.

val rpcs : unit -> Rpc.Any.t list

All rpcs supported by dispatch_multi.

val versions : unit -> Core_kernel.Int.Set.t

All versions supported by dispatch_multi (useful for computing which old versions may be pruned).

val name : string